requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.12.2
requests>=2.28.0
pandas>=1.5.0
firecrawl-py>=0.0.16
ics>=0.7.2
pylatex>=1.4.1
python-dotenv>=0.20.0
Flask>=3.0.0
gunicorn>=21.2.0
reportlab>=4.1.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
